Reintroduce mad (append), requested by a user
This essentially reverts commits 8901464, e57a2ca and 38ab87c.

As well, remove the ZAPPEND flag, which only applied to MS/DOS, and add
.gitignore patterns for executables ending in .exe.

+++ mmv.c
@@ -78,6 +78,7 @@
 #define NORMMOVE 0x008
 #define XMOVE 0x010
 #define DIRMOVE 0x020
+#define APPEND 0x040
 #define HARDLINK 0x100
 #define SYMLINK 0x200
 
@@ -86,6 +87,7 @@
 #define LINK (HARDLINK | SYMLINK)
 
 static char COPYNAME[] = "mcp";
+static char APPENDNAME[] = "mad";
 static char LINKNAME[] = "mln";
 
 #define ASKDEL 0
@@ -865,7 +867,7 @@ static int badrep(HANDLE *hfrom, FILEINFO *ffrom, HANDLE **phto, char **pnto, FI
 	if ((ffrom->fi_stflags & FI_LINKERR) && !(op & (MOVE | SYMLINK)))
 		printf("%s -> %s : source file is a badly aimed symbolic link.\n",
 			pathbuf, fullrep);
-	else if ((op & COPY) && access(pathbuf, R_OK))
+	else if ((op & (COPY | APPEND)) && access(pathbuf, R_OK))
 		printf("%s -> %s : no read permission for source file.\n",
 			pathbuf, fullrep);
 	else if (
@@ -1408,7 +1410,7 @@ static int baddel(REP *p)
 	char *t = fto->fi_name, *f = p->r_ffrom->fi_name;
 	char *hnf = hfrom->h_name, *hnt = hto->h_name;
 
-	if (delstyle == NODEL && !(p->r_flags & R_DELOK))
+	if (delstyle == NODEL && !(p->r_flags & R_DELOK) && !(op & APPEND))
 		printf("%s%s -> %s%s : old %s%s would have to be %s.\n",
 			hnf, f, hnt, t, hnt, t,
 			(op & OVERWRITE) ? "overwritten" : "deleted");
@@ -1419,12 +1421,12 @@ static int baddel(REP *p)
 		fto->fi_stflags & FI_ISDIR
 	)
 		printf("%s%s -> %s%s : %s%s%s is a directory.\n",
-			hnf, f, hnt, t, "old ", hnt, t);
-	else if ((fto->fi_stflags & FI_NODEL) && !(op & OVERWRITE))
+		       hnf, f, hnt, t, (op & APPEND) ? "" : "old ", hnt, t);
+	else if ((fto->fi_stflags & FI_NODEL) && !(op & (APPEND | OVERWRITE)))
 		printf("%s%s -> %s%s : old %s%s lacks delete permission.\n",
 			hnf, f, hnt, t, hnt, t);
 	else if (
-		(op & OVERWRITE) &&
+		(op & (APPEND | OVERWRITE)) &&
 		!fwritable(hnt, fto)
 	) {
 		printf("%s%s -> %s%s : %s%s %s.\n",
@@ -1471,7 +1473,7 @@ static void showdone(REP *fin)
 				p->r_flags & R_ISALIASED ? '=' : '-',
 				p->r_flags & R_ISCYCLE ? '^' : '>',
 				p->r_hto->h_name, p->r_nto,
-				(p->r_fdel != NULL) ? " (*)" : "");
+				(p->r_fdel != NULL && !(op & APPEND)) ? " (*)" : "");
 		}
 }
 
@@ -1487,6 +1489,22 @@ static int snap(REP *first, REP *p)
 	printf("The following left undone:\n");
 	noex = 1;
 	return(first != p);
+}
+
+static long appendalias(REP *first, REP *p, int *pprintaliased)
+{
+	long ret = 0l;
+
+	struct stat fstat;
+
+	if (stat(fullrep, &fstat)) {
+		fprintf(stderr, "append cycle stat on %s has failed.\n", fullrep);
+		*pprintaliased = snap(first, p);
+	}
+	else
+		ret = fstat.st_size;
+
+	return(ret);
 }
 
 static int movealias(REP *first, REP *p, int *pprintaliased)
@@ -1515,7 +1533,7 @@ static int movealias(REP *first, REP *p, int *pprintaliased)
 #define IRWMASK (S_IRUSR | S_IWUSR)
 #define RWMASK (IRWMASK | (IRWMASK >> 3) | (IRWMASK >> 6))
 
-static int copy(FILEINFO *ff)
+static int copy(FILEINFO *ff, off_t len)
 {
 	char buf[BUFSIZ];
 	int f, t, mode;
@@ -1526,19 +1544,32 @@ static int copy(FILEINFO *ff)
 
 	if ((f = open(pathbuf, O_RDONLY | O_BINARY, 0)) < 0)
 		return(-1);
-	perm = (op & OVERWRITE) ?
+	perm = (op & (APPEND | OVERWRITE)) ?
 		(~oldumask & RWMASK) | (ff->fi_mode & (mode_t)~RWMASK) :
 		ff->fi_mode;
 
-	mode = O_CREAT | O_TRUNC | O_WRONLY;
+	mode = O_CREAT | (op & APPEND ? 0 : O_TRUNC) | O_WRONLY;
 	t = open(fullrep, mode, perm);
 	if (t < 0) {
 		close(f);
 		return(-1);
 	}
-	while ((k = read(f, buf, BUFSIZ)) > 0 && write(t, buf, (size_t)k) == k)
-		;
-	if (!(op & OVERWRITE))
+	if (op & APPEND)
+		lseek(t, (off_t)0, SEEK_END);
+	if ((op & APPEND) && len != (off_t)-1) {
+		while (
+			len != 0 &&
+			(k = read(f, buf, (len > BUFSIZ) ? BUFSIZ : (size_t)len)) > 0 &&
+			write(t, buf, (size_t)k) == k
+		)
+			len -= k;
+		if (len == 0)
+			k = 0;
+	}
+	else
+		while ((k = read(f, buf, BUFSIZ)) > 0 && write(t, buf, (size_t)k) == k)
+			;
+	if (!(op & (APPEND | OVERWRITE)))
 		if (
 			stat(pathbuf, &fstat) ||
 			(
@@ -1553,7 +1584,8 @@ static int copy(FILEINFO *ff)
 	close(f);
 	close(t);
 	if (k != 0) {
-		unlink(fullrep);
+		if (!(op & APPEND))
+			unlink(fullrep);
 		return(-1);
 	}
 	return(0);
@@ -1570,7 +1602,7 @@ static int myunlink(char *n)
 
 static int copymove(REP *p)
 {
-	return(copy(p->r_ffrom) || myunlink(pathbuf));
+	return(copy(p->r_ffrom, -1L) || myunlink(pathbuf));
 }
 
 static void doreps(void)
@@ -1579,6 +1611,7 @@ static void doreps(void)
 	unsigned k;
 	int printaliased = 0, alias = 0;
 	REP *first, *p;
+	long aliaslen = 0l;
 
 	signal(SIGINT, breakrep);
 
@@ -1600,20 +1633,25 @@ static void doreps(void)
 				p->r_hto->h_di->di_flags &= ~DI_NONEXISTENT;
 			}
 			strcat(fullrep, p->r_nto);
-			if (!noex && (p->r_flags & R_ISCYCLE))
-				alias = movealias(first, p, &printaliased);
+			if (!noex && (p->r_flags & R_ISCYCLE)) {
+				if (op & APPEND)
+					aliaslen = appendalias(first, p, &printaliased);
+				else
+					alias = movealias(first, p, &printaliased);
+			}
 			strcpy(pathbuf, p->r_hfrom->h_name);
 			fstart = pathbuf + strlen(pathbuf);
-			if (p->r_flags & R_ISALIASED)
+			if ((p->r_flags & R_ISALIASED) && !(op & APPEND))
 				sprintf(fstart, "%s%03d", TEMP, alias);
 			else
 				strcpy(fstart, p->r_ffrom->fi_name);
 			if (!noex) {
-				if (p->r_fdel != NULL && !(op & OVERWRITE))
+				if (p->r_fdel != NULL && !(op & (APPEND | OVERWRITE)))
 					myunlink(fullrep);
 				if (
-					(op & COPY) ?
-						copy(p->r_ffrom) :
+					(op & (COPY | APPEND)) ?
+						copy(p->r_ffrom,
+							p->r_flags & R_ISALIASED ? aliaslen : -1L) :
 					(op & HARDLINK) ?
 						link(pathbuf, fullrep) :
 					(op & SYMLINK) ?
@@ -1637,7 +1675,7 @@ static void doreps(void)
 					p->r_flags & R_ISALIASED ? '=' : '-',
 					p->r_flags & R_ISCYCLE ? '^' : '>',
 					fullrep,
-					(p->r_fdel != NULL) ? " (*)" : "",
+					(p->r_fdel != NULL && !(op & APPEND)) ? " (*)" : "",
 					noex ? "" : " : done");
 			}
 		}
@@ -1711,6 +1749,8 @@ int main(int argc, char *argv[])
 		op = NORMCOPY;
 	else if (args_info.overwrite_given != 0)
 		op = OVERWRITE;
+	else if (args_info.append_given != 0)
+		op = APPEND;
 	else if (args_info.hardlink_given != 0)
 		op = HARDLINK;
 	else if (args_info.symlink_given != 0)
@@ -1718,6 +1758,8 @@ int main(int argc, char *argv[])
 	else {
 		if (strcmp(program_name, COPYNAME) == 0)
 			op = NORMCOPY;
+		else if (strcmp(program_name, APPENDNAME) == 0)
+			op = APPEND;
 		else if (strcmp(program_name, LINKNAME) == 0)
 			op = HARDLINK;
 		else
@@ -1747,13 +1789,14 @@ int main(int argc, char *argv[])
 	}
 
 	domatch(frompat, topat);
-	checkcollisions();
+	if (!(op & APPEND))
+		checkcollisions();
 	findorder();
 	if (op & (COPY | LINK))
 		nochains();
 	scandeletes(baddel);
 	goonordie();
-	if (delstyle == ASKDEL)
+	if (!(op & APPEND) && delstyle == ASKDEL)
 		scandeletes(skipdel);
 	doreps();
